Add tests for empty contract
